+++ test-app/src/domain/test/english/syntax.feature.ts
@@ -1,5 +1,5 @@
 import { expect } from "vitest";
-import { Given } from "vitest-bdd";
+import { Given } from "@epure/vitest";
 
 Given("I count {number} time(s)", ({ Then }, number: number) => {
   Then("I count {number} time(s)", (count: number) => {